Ignore meta-data tokens inside code blocks in message
Any tokens inside code blocks will not be used when parsing meta-data from a message. i.e you can @ mention someone or add a hashtag in a code block and it will not be added to the meta data of the message card.

Note - this does not affect the rendering of these tokens. This will be addressed in a separate PR.

Change-type: patch
Signed-off-by: Graham McCulloch <graham@balena.io>