Update rendition from v21.x to v25.x
The breaking changes in rendition apply only to the `AutoUI` and
`DropDownButton` components, which aren't part of the components
exported by this module. However, because the peer dependency has now
increased, this new version will no longer be compatible with a number
of implementing projects, so this is a breaking change.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>